Code Refactoring: Split widget code
