+++ lib/widgets/selector_lists/ccitt2_selection.dart
@@ -4,7 +4,7 @@ import 'package:gc_wizard/widgets/common/gcw_toollist.dart';
 import 'package:gc_wizard/widgets/registry.dart';
 import 'package:gc_wizard/widgets/selector_lists/gcw_selection.dart';
 import 'package:gc_wizard/widgets/tools/crypto_and_encodings/ccitt2.dart';
-import 'package:gc_wizard/widgets/tools/crypto_and_encodings/symbol_table.dart';
+import 'package:gc_wizard/widgets/tools/symbol_tables/symbol_table.dart';
 import 'package:gc_wizard/widgets/utils/common_widget_utils.dart';
 
 class CCITT2Selection extends GCWSelection {
